fix: 우승자 객체 추가
